[es6-bookreader-core/BookModel] create BookModel.js
